+++ HISTORY.rst
@@ -9,7 +9,7 @@ History
 * Fix issues #108 #109 and merges #110 #120
 * Remove requirements.txt, add docs/rtd-pip-requirements, update .travis.yaml
 * Add support for running at sites where data is locally visible (e.g., nersc)
-* Add support for reading spFlat files
+* Add support for reading spArc and spFlat calibration files.
 
 0.3.0 (2017-12-17)
 ------------------